+++ src/Utils/closest/index.js
@@ -1,0 +1,3 @@
+import closest from './closest';
+
+export default closest;
