+++ src/Utils/index.js
@@ -1,0 +1,7 @@
+import closest from './closest';
+import scroll from './scroll';
+
+export {
+  closest,
+  scroll,
+};
