+++ src/Utils/scroll/index.js
@@ -1,0 +1,3 @@
+import scroll from './scroll';
+
+export default scroll;
